03-microservices-scaling/api-gateway/requirements.txt
flask
gunicorn
python-dotenv
requests
pyjwt
flask-cors